batchkey gen in views.py
Tested on randomly generated lists of id, size up to 10000. Runtime is O(n!), but in real use cases of input size <1000, my machine's execution time is less than 1 second. For now, O(n!) is realistically fine, but still disgusting.

Also, this is essentially not integrated to the webapp (discussed during meeting, in regards to HTML forms, posts, etc)

The key generator function, make_id_dictionary, is in views.py.

 It can be easily put where it needs to be once session functionality integrations begins, though!