Skeletal attendance list compiler done
Takes in a list of students and the number of connections they have made (easily modified depending on DB structure).

Checks each student to see if they have met the attendance quota fully, partially, or have zero connections, then sorts them into respective lists.

Final returned list is a list of lists:
[[not present], [partially present[, [fully present]]

In future, presence may also take in time of connection as a parameter (for students being making connections towards the end of a session, for instance)